lossless-claw-5a9.4: isolate raw-id replay filter
